requirements.txt
rich
paramiko
python-whois
requests
